[tf] add a way to actually get the full line passed into a macro, not just the post-argv parsed positional args
Calling a function directly just populates the arguments directly, but
calling a macro - or it being called from say a trigger/hook/etc -
will invoke the positional argument parser.

This has the fun side effect of stripping whitespace from between
"arguments" with no way to get the original line back.

This is especially annoying if you're trying to say, use a full line
glob match, or handle input text that has multiple spaces in between
for formatting.

So, this adds ${A} - populated with the full line that was passed
into do_macro() and /before/ all the parsing is done to it.

It should only be used in things called as macros, as calling
functions doesn't touch argstring and thus ${A}.  I'd rather not
have to fix that. :-)